Don't buffer cover commands
For common UP/DOWN/STOP commands, it just makes the device seem
very laggy. If your device supports position, and you're actually
using a UI that continuously sends position commands as you're
dragging a slider around, it should be the responsibility of the
device to handle that gracefully (either by throttling them itself,
or not restarting movement when the new command continues movement
in the same direction). Given ESPHome is all over TCP, a high
rate of commands shouldn't be a connection issue.

Signed-off-by: Cody Cutrer <cody@cutrer.us>